+++ crates/miden-testing/src/kernel_tests/mod.rs
@@ -9,6 +9,7 @@ use ::assembly::{
     LibraryPath,
     ast::{Module, ModuleKind},
 };
+use anyhow::Context;
 use assert_matches::assert_matches;
 use miden_lib::{
     note::{create_p2id_note, create_p2idr_note},
@@ -16,7 +17,7 @@ use miden_lib::{
     utils::word_to_masm_push_string,
 };
 use miden_objects::{
-    Felt, FieldElement, MIN_PROOF_SECURITY_LEVEL, Word,
+    Felt, FieldElement, Hasher, MIN_PROOF_SECURITY_LEVEL, TransactionScriptError, Word,
     account::{Account, AccountBuilder, AccountComponent, AccountId, AccountStorage, StorageSlot},
     assembly::DefaultSourceManager,
     asset::{Asset, AssetVault, FungibleAsset, NonFungibleAsset},
@@ -817,22 +818,22 @@ fn prove_witness_and_verify() {
 // ================================================================================================
 
 #[test]
-fn test_tx_script() {
+fn test_tx_script_inputs() {
     let tx_script_input_key = [Felt::new(9999), Felt::new(8888), Felt::new(9999), Felt::new(8888)];
     let tx_script_input_value = [Felt::new(9), Felt::new(8), Felt::new(7), Felt::new(6)];
     let tx_script_src = format!(
         "
-    begin
-        # push the tx script input key onto the stack
-        push.{key}
+        begin
+            # push the tx script input key onto the stack
+            push.{key}
 
-        # load the tx script input value from the map and read it onto the stack
-        adv.push_mapval adv_loadw
+            # load the tx script input value from the map and read it onto the stack
+            adv.push_mapval adv_loadw
 
-        # assert that the value is correct
-        push.{value} assert_eqw
-    end
-",
+            # assert that the value is correct
+            push.{value} assert_eqw
+        end
+        ",
         key = word_to_masm_push_string(&tx_script_input_key),
         value = word_to_masm_push_string(&tx_script_input_value)
     );
@@ -845,7 +846,6 @@ fn test_tx_script() {
     .unwrap();
 
     let tx_context = TransactionContextBuilder::with_standard_account(ONE)
-        .with_mock_notes_preserved()
         .tx_script(tx_script)
         .build();
 
@@ -856,6 +856,96 @@ fn test_tx_script() {
         "Transaction execution failed {:?}",
         executed_transaction,
     );
+}
+
+#[test]
+fn test_tx_script_args() -> anyhow::Result<()> {
+    let tx_script_src = r#"
+        begin
+            # => [TX_SCRIPT_ARGS_KEY]
+            # `TX_SCRIPT_ARGS_KEY` value, which is located on the stack at the beginning of 
+            # the execution, is the advice map key which allows to obtain transaction script args 
+            # which were specified during the `TransactionScript` creation (see below). In this
+            # example transaction args is an array of Felts from 1 to 7.
+
+            # move the transaction args from advice map to the advice stack
+            adv.push_mapval
+            # OS => [TX_SCRIPT_ARGS_KEY]
+            # AS => [7, 6, 5, 4, 3, 2, 1]
+
+            # drop the args commitment
+            dropw
+            # OS => []
+            # AS => [7, 6, 5, 4, 3, 2, 1]
+
+            # Move the transaction arguments array from advice stack to the operand stack. It 
+            # consists of 7 Felts, so we could use `adv_push.7` instruction to load them all at once
+            adv_push.7
+            # OS => [7, 6, 5, 4, 3, 2, 1]
+            # AS => []
+
+            # assert that the transaction arguments array consists of correct values
+            push.4.5.6.7
+            assert_eqw.err="last four values in the transaction args array are incorrect"
+            # OS => [3, 2, 1]
+            # AS => []
+            
+            # To use more convenient `assert_eqw` instruction we should push `0` first, but notice 
+            # that there are only three values from the original array left on the stack. We could 
+            # do so because there are no other elements on the stack left except for the argument 
+            # values (hidden values deeper on the stack are zeros). In case there are some values 
+            # deeper on the stack, we should assert values one by one using `push.n assert_eq`.
+            push.0.1.2.3
+            assert_eqw.err="first three values in the transaction args array are incorrect"
+        end"#;
+
+    let tx_script =
+        TransactionScript::compile(tx_script_src, [], TransactionKernel::testing_assembler())
+            .context("failed to compile transaction script")?
+            .with_args(&[
+                ONE,
+                Felt::new(2),
+                Felt::new(3),
+                Felt::new(4),
+                Felt::new(5),
+                Felt::new(6),
+                Felt::new(7),
+            ])?;
+
+    let tx_context = TransactionContextBuilder::with_standard_account(ONE)
+        .tx_script(tx_script)
+        .build();
+
+    tx_context.execute().context("failed to execute transaction")?;
+
+    Ok(())
+}
+
+#[test]
+fn test_tx_script_args_collision() -> anyhow::Result<()> {
+    let collision_elements = vec![ONE, Felt::new(2), Felt::new(3), Felt::new(4)];
+    let collision_key = Hasher::hash_elements(&collision_elements);
+
+    let script_args_collision_err = TransactionScript::compile(
+        "begin nop end",
+        [(*collision_key, vec![ONE, Felt::new(2)])],
+        TransactionKernel::testing_assembler(),
+    )
+    .context("failed to compile transaction script")?
+    .with_args(&collision_elements)
+    .unwrap_err();
+
+    assert_matches!(script_args_collision_err, TransactionScriptError::ScriptArgsCollision {
+        key,
+        new_value,
+        old_value,
+    } => {
+        assert_eq!(key, collision_key);
+        assert_eq!(new_value, collision_elements);
+        assert_eq!(old_value, [ONE, Felt::new(2)]);
+    });
+
+    Ok(())
 }
 
 /// Tests that an account can call code in a custom library when loading that library into the